+++ MasterThesis/RePair_v2/MyTest.h
@@ -7,17 +7,17 @@ public:
 
 	int MyTest::SanityCheck(
 		vector<SymbolRecord*> & sequenceArray,
-		vector<PairRecord*>& priorityQueue, unordered_map<unsigned int,
-		unordered_map<unsigned int, PairTracker >> &activePairs);
+		vector<PairRecord*>& priorityQueue, unordered_map<long,
+		unordered_map<long, PairTracker >> &activePairs);
 	string MyTest::SanityCheckThreadingPointersDetailed(vector<SymbolRecord*> & sequenceArray);
 	string MyTest::SanityCheckPairRecordsDetailed(vector<SymbolRecord*> & sequenceArray, vector<PairRecord*>& priorityQueue);
 	int MyTest::SanityCheckThreadingPointers(vector<SymbolRecord*> & sequenceArray);
-	int MyTest::SanityCheckPairRecords(vector<SymbolRecord*> & sequenceArray, vector<PairRecord*>& priorityQueue, unordered_map<unsigned int, unordered_map<unsigned int, PairTracker>>& activePairs);
+	int MyTest::SanityCheckPairRecords(vector<SymbolRecord*> & sequenceArray, vector<PairRecord*>& priorityQueue, unordered_map<long, unordered_map<long, PairTracker>>& activePairs);
 	void MyTest::buildSequenceArray(vector<SymbolRecord*> & sequenceArray, long numbers[], int count);
 	void MyTest::charArrayToSequence(vector<SymbolRecord*> &  sequence, string input, int size);
 	string MyTest::SequenceToString(vector<SymbolRecord*> & sequenceArray);
 	string MyTest::SequenceToCompleteString(vector<SymbolRecord*> & sequenceArray);
-	void MyTest::ActivePairs(string msg, unordered_map<unsigned int, unordered_map<unsigned int, PairTracker>>& activePairs);
+	void MyTest::ActivePairs(string msg, unordered_map<long, unordered_map<long, PairTracker>>& activePairs);
 	bool MyTest::inPriorityQueue(
 		PairRecord * pairRecord,
 		vector<PairRecord*> & priorityQueue);
