+++ MasterThesis/RePair_v2/RePair_v2.cpp
@@ -13,11 +13,11 @@ int main(int argc, char* argv[])
 	int blockSize;
 	blockSize = 1048576;
 
-	unordered_map<unsigned int, unordered_map<unsigned int, PairTracker>> activePairs;
+	unordered_map<long, unordered_map<long, PairTracker>> activePairs;
 	vector<SymbolRecord*> sequenceArray;
 	vector<PairRecord*> priorityQueue;
-	unordered_map<unsigned int, Pair> dictionary;
-	unsigned int symbols(initialSymbolValue);
+	unordered_map<long, Pair> dictionary;
+	long symbols(initialSymbolValue);
 
 	Algorithm algorithm;
 	AlgorithmP algP;
